(adjust) PrimaryKeyValueGenerator 로 PK 값 생성 책임 분리
