+++ src/main/java/persistence/entity/persistencecontext/Snapshot.java
@@ -1,9 +1,9 @@
 package persistence.entity.persistencecontext;
 
-import persistence.sql.ddl.PrimaryKeyClause;
-
 import java.util.HashMap;
 import java.util.Map;
+
+import static persistence.entity.generator.PrimaryKeyValueGenerator.primaryKeyValue;
 
 public class Snapshot {
     private final Map<EntityKey, Object> snapshot;
@@ -14,7 +14,7 @@ public class Snapshot {
 
     public void put(Object entity) {
         var clazz = entity.getClass();
-        snapshot.put(new EntityKey(clazz, PrimaryKeyClause.primaryKeyValue(entity)), entity);
+        snapshot.put(new EntityKey(clazz, primaryKeyValue(entity)), entity);
     }
 
     public void remove(Object entity) {
