+++ src/main/java/persistence/sql/ddl/PrimaryKeyClause.java
@@ -3,12 +3,9 @@ package persistence.sql.ddl;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import persistence.entity.exception.InvalidPrimaryKeyException;
-import persistence.entity.exception.PrimaryKeyEditingNotAllowedException;
 import persistence.sql.exception.NotIdException;
 
 import java.lang.reflect.Field;
-import java.util.Arrays;
 import java.util.Map;
 
 public class PrimaryKeyClause {
@@ -30,20 +27,6 @@ public class PrimaryKeyClause {
         this.name = field.getName();
         this.dataType = field.getType().getSimpleName();
         this.generationType = getType(field);
-    }
-
-    public static Long primaryKeyValue(Object entity) {
-        Field idField = Arrays.stream(entity.getClass().getDeclaredFields())
-                .filter(x -> x.isAnnotationPresent(Id.class))
-                .findAny()
-                .orElseThrow(InvalidPrimaryKeyException::new);
-
-        idField.setAccessible(true);
-        try {
-            return (Long) idField.get(entity);
-        } catch (IllegalAccessException e) {
-            throw new InvalidPrimaryKeyException();
-        }
     }
 
     private static GenerationType getType(Field field) {
